# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Buildings History API Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/building-part/": {
            "get": {
                "tags": [
                    "Building Part"
                ],
                "operationId": "a21d65d5903dffb016a1a40a5af1194e",
                "responses": {
                    "200": {
                        "description": "List of building part",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/building-part/nearest": {
            "get": {
                "tags": [
                    "Building Part"
                ],
                "operationId": "5c971b965f7c7535574e73e975cae986",
                "parameters": [
                    {
                        "name": "latitude",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    },
                    {
                        "name": "longitude",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    },
                    {
                        "name": "distance",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    },
                    {
                        "name": "imagedirection",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get nearest building part",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/codepoint": {
            "get": {
                "tags": [
                    "Codepoint"
                ],
                "operationId": "e500a77e7fede51e4068f294ea35a15b",
                "parameters": [
                    {
                        "name": "postcode",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "BA1 0AH"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of codepoint",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/galleries/": {
            "get": {
                "tags": [
                    "Gallery"
                ],
                "operationId": "07469ec7f7e00f9b1b1e09087f6e04fd",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "slug",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "created_at",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get all galleries",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/galleries/to-brick-collections-format": {
            "get": {
                "tags": [
                    "Gallery"
                ],
                "operationId": "4ed4f024f5b0f0eb9a3a8122312dbf0f",
                "responses": {
                    "200": {
                        "description": "Get all galleries and show to_brick collections format",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/galleries/sync": {
            "get": {
                "tags": [
                    "Gallery"
                ],
                "operationId": "39f8843d235319bd1261129c23f93c50",
                "responses": {
                    "200": {
                        "description": "Sync all galleries with brick_by_brick collections",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/images/": {
            "get": {
                "tags": [
                    "Image"
                ],
                "operationId": "6eedfbb0e4173adcf6920fa20758e31a",
                "parameters": [
                    {
                        "name": "gallery_id",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filename",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "long_description",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "exif_data_latitude",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "exif_data_longitude",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "exif_data_altitude",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "exif_data_focal_length",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "exif_data_iso",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "exif_data_taken_at",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "exif_data_gps_img_direction",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "exif_data_gps_longitude_ref",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "exif_data_gps_latitude_ref",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "created_at",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get all images",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/images/to-brick-items-format": {
            "get": {
                "tags": [
                    "Image"
                ],
                "operationId": "2be9f4f649b4da5a185872aca9772925",
                "parameters": [
                    {
                        "name": "gallery_id",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get all images and show to_brick items format",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/images/sync": {
            "get": {
                "tags": [
                    "Image"
                ],
                "operationId": "4d104e1ace696d0b09fc2b0ddb5e22ff",
                "responses": {
                    "200": {
                        "description": "Sync all images to brick_by_brick items",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/geo/": {
            "get": {
                "tags": [
                    "Geo"
                ],
                "operationId": "fbc4f5efbcc1ca91188ce22ae1ae592f",
                "parameters": [
                    {
                        "name": "minEasting",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    },
                    {
                        "name": "minNorthing",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    },
                    {
                        "name": "maxEasting",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    },
                    {
                        "name": "maxNorthing",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get all geos",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Geo"
                ],
                "operationId": "3aae36eb42c17a49c29fd234fde622ad",
                "parameters": [
                    {
                        "name": "geom",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "osid",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "toid",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "height_max",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "symbol",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post Geo",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/geo/nearest": {
            "get": {
                "tags": [
                    "Geo"
                ],
                "operationId": "79a61eb6abd4b6e80b5c374eb4ae0a28",
                "parameters": [
                    {
                        "name": "latitude",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    },
                    {
                        "name": "longitude",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    },
                    {
                        "name": "radius",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get nearest geos with radius",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/geo/upload/": {
            "post": {
                "tags": [
                    "Geo"
                ],
                "operationId": "008a9a5d8eb2933ebf6d45ccbbd66f50",
                "parameters": [
                    {
                        "name": "geom",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "osid",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "toid",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "height_max",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "symbol",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Upload Geo",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/sx-data/": {
            "get": {
                "tags": [
                    "SX Data"
                ],
                "operationId": "b52932c8edf787a6b73ddb659710b838",
                "responses": {
                    "200": {
                        "description": "Get all SX Data",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/layer-sx/": {
            "get": {
                "tags": [
                    "Topography Layer SX"
                ],
                "operationId": "64b14c5b8a1783fab44e56ce0a3146f5",
                "responses": {
                    "200": {
                        "description": "Get all Topography Layer SX Data",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/layer-sx/nearest": {
            "get": {
                "tags": [
                    "Topography Layer SX"
                ],
                "operationId": "33240adce5d46b67a7e05baa2e40cd28",
                "parameters": [
                    {
                        "name": "latitude",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    },
                    {
                        "name": "longitude",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    },
                    {
                        "name": "radius",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get nearest SX with radius",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/uprn": {
            "get": {
                "tags": [
                    "UPRN"
                ],
                "operationId": "ac2dbe4d1e6b3b5873f36c8bb1e0c2b3",
                "parameters": [
                    {
                        "name": "uprn",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "1"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of UPRN address",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/ward-south-east/": {
            "get": {
                "tags": [
                    "Ward South East"
                ],
                "operationId": "9c503b0b297d5316f520462c643a60bd",
                "parameters": [
                    {
                        "name": "minLat",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    },
                    {
                        "name": "minLng",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    },
                    {
                        "name": "maxLat",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    },
                    {
                        "name": "maxLng",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of Ward South East",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Building Part",
            "description": "Building Part"
        },
        {
            "name": "Codepoint",
            "description": "Codepoint"
        },
        {
            "name": "Gallery",
            "description": "Gallery"
        },
        {
            "name": "Image",
            "description": "Image"
        },
        {
            "name": "Geo",
            "description": "Geo"
        },
        {
            "name": "SX Data",
            "description": "SX Data"
        },
        {
            "name": "Topography Layer SX",
            "description": "Topography Layer SX"
        },
        {
            "name": "UPRN",
            "description": "UPRN"
        },
        {
            "name": "Ward South East",
            "description": "Ward South East"
        }
    ]
}